fix: Adjusted group with normalizer
