+++ src/lib/wapi/validation/group.js
@@ -1,5 +1,6 @@
+import { normalizePhoneNumber } from '../helper/normalize-phone-number'
+
 const { GROUP_ERRORS } = require('../constants/group-errors')
-const phoneUtil = require('google-libphonenumber').PhoneNumberUtil.getInstance()
 
 export function verifyGroup(chat, needsToBeAdmin = false) {
   if (!chat || chat.erro || chat.status === 404 || !chat.isGroup) {
@@ -92,36 +93,13 @@ export const getAddParticipantStatusError = (statusError) => {
   }
 }
 
-export const normalizePhoneNumber = (contactPhoneNumber) => {
-  const phoneNumber = contactPhoneNumber.replace('@c.us', '')
-  const phoneCountry = getCountryByPhoneNumber(phoneNumber)
+export const getContactIndex = (phoneNumber, contacts) => {
+  const normalizedPhoneNumber = normalizePhoneNumber(phoneNumber)
+  return contacts.findIndex((contact) => {
+    const normalizedContactPhoneNumber = normalizePhoneNumber(
+      contact.phoneNumber
+    )
 
-  switch (phoneCountry) {
-    case 'BR':
-      if (hasThirteenDigits(phoneNumber)) {
-        contactPhoneNumber = removeNineDigitInPhoneNumber(phoneNumber) + '@c.us'
-      }
-      break
-  }
-
-  return contactPhoneNumber
-}
-
-function removeNineDigitInPhoneNumber(phoneNumber) {
-  return (
-    phoneNumber.toString().substr(0, 4) + '' + phoneNumber.toString().substr(5)
-  )
-}
-
-function hasThirteenDigits(phoneNumber) {
-  return phoneNumber.toString().length === 13
-}
-
-function getCountryByPhoneNumber(phoneNumber) {
-  const phone = getPhoneParse(phoneNumber)
-  return phoneUtil.getRegionCodeForNumber(phone)
-}
-
-function getPhoneParse(phoneNumber) {
-  return phoneUtil.parse(`+${phoneNumber}`)
+    return normalizedContactPhoneNumber === normalizedPhoneNumber
+  })
 }
